requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-redis==4.10.0
django-reversion==3.0.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
future==0.17.1
httplib2==0.12.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mysqlclient==1.3.13
Naked==0.1.31
Pillow==5.3.0
PyJWT==1.7.0
pytz==2018.7
PyYAML==3.13
redis==3.0.1
requests==2.20.1
six==1.11.0
status==0.2.5
uritemplate==3.0.0
urllib3==1.24.1
XlsxWriter==1.1.2
xlwt==1.3.0
